+++ app/routes/static.py
@@ -1,11 +1,10 @@
 from fastapi import APIRouter, Depends
 
-from app.helpers.oauth import TokenService
+from app.helpers.jwt_token import get_current_user
 
 router = APIRouter()
 
-current_user_validator = TokenService.get_current_user
 
 @router.get("/protected")
-async def protected(token: str = Depends(current_user_validator)):
+async def protected(token: str = Depends(get_current_user)):
     return {"message": "This is a protected endpoint", "token": token}
